backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pytesseract
langchain
langchain-core
langchain-openai
langgraph
sqlalchemy
python-dotenv
httpx
openai
